+++ Source/EventFlow.MongoDB/ReadStores/Attributes/MongoDbGeoSpatialIndexAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace EventFlow.MongoDB.ReadStores.Attributes
+{
+    [AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
+    public class MongoDbGeoSpatialIndexAttribute : Attribute
+    {
+
+    }
+}